Change config file format from INI to TOML
This simplifies parsing (because everything is now parsed with Serde)
and allows for more data types in the config file (like lists and maps).
It also reduces the total dependency count by 10.

The default filename was changed from `config.ini` to `config.toml`.
Since TOML is a superset of INI, the existing config should remain
valid.