Migrate to Tokio 1.0 / async-await
This is a major rewrite which moves the codebase to hyper 0.14 and thus
to Tokio 1.0. This means we can finally make use of async-await syntax.
This simplifies a lot of things and makes the codebase much easier to
read and maintain!

The project temporarily depends on a git version of the a2 crate. There
will be a new a2 release as soon as the changes are tested in production.